Read numberOfColumns from config instead of calculating (#46)
* NUMBER_OF_COLUMNS is now in config, better to read it
* removed hardcode from fieldElementsPerCell as there is no more division by it

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/Spec.java
@@ -956,7 +956,7 @@ public class Spec {
     return blobSidecar.getIndex().mod(specConfigDeneb.getBlobSidecarSubnetCount());
   }
 
-  public Optional<UInt64> getNumberOfDataColumns() {
+  public Optional<Integer> getNumberOfDataColumns() {
     return getSpecConfigEip7594().map(SpecConfigEip7594::getNumberOfColumns);
   }
